Mean Shift Clustering Learner

A LearnerClust for Mean Shift clustering implemented in LPCM::ms(). There is no predict method for LPCM::ms(), so the method returns cluster labels for the 'training' data.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("clust.meanshift")
lrn("clust.meanshift")

Meta Information

Task type: “clust”

Predict Types: “partition”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3cluster, LPCM

Parameters

Id	Type	Default	Range

h	untyped	-	-

subset	untyped	-	-

scaled	integer	1	\([0, \infty)\)

iter	integer	200	\([1, \infty)\)

thr	numeric	0.01	\((-\infty, \infty)\)

References

Cheng, Yizong (1995). “Mean shift, mode seeking, and clustering.” IEEE transactions on pattern analysis and machine intelligence, 17(8), 790–799.

See also

Super classes

mlr3::Learner -> mlr3cluster::LearnerClust -> LearnerClustMeanShift

Methods

Public methods

LearnerClustMeanShift$new()

LearnerClustMeanShift$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$train()

mlr3cluster::LearnerClust$reset()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClustMeanShift$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClustMeanShift$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("LPCM")) {
  learner = mlr3::lrn("clust.meanshift")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
#> <LearnerClustMeanShift:clust.meanshift>: Mean Shift Clustering
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3cluster, LPCM
#> * Predict Types:  [partition]
#> * Feature Types: logical, integer, numeric
#> * Properties: complete, exclusive, partitional
#> [1] "h"      "subset" "scaled" "iter"   "thr"