Dunn2 Index

An alternative formulation of the Dunn index that uses average distances instead of extremes. It is defined as the ratio of the minimum average between-cluster distance to the maximum average within-cluster distance: \(D_2 = \min_{i \neq j} \bar{d}(C_i, C_j) / \max_k \bar{d}(C_k)\). This variant is more robust to outliers than the standard Dunn index. Higher values indicate better separation.

Details

If the task contains factor or ordered features, Gower distances (cluster::daisy()) are used instead of Euclidean distances.

Dictionary

This mlr3::Measure can be instantiated via the dictionary mlr3::mlr_measures or with the associated sugar function mlr3::msr():

mlr_measures$get("clust.dunn2")
msr("clust.dunn2")

Meta Information

Task type: “clust”

Range: \([0, \infty)\)

Minimize: FALSE

Average: macro

Required Prediction: “partition”

Required Packages: mlr3, mlr3cluster, cluster

References

Dunn, C J (1974). “Well-separated clusters and optimal fuzzy partitions.” Journal of Cybernetics, 4(1), 95–104. doi:10.1080/01969727408546059 .

See also